Add `for` to label tags
In the origami change the labels lost their for attributes as the label
now wraps the input. This ensures that all label tags have a appropriate
for attribute which points at the input they are targeting. Having both
a for attribute and wrapping the input is valid.

This was primarily picked up due to the end to end tests trying to find
the input for a radio button to click.